requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.1.0
groq==0.5.0
httpx==0.25.2
sentence-transformers
onnxruntime==1.24.1
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
slowapi==0.1.9
reportlab==4.0.7